+++ src/estd/internal/string.h
@@ -117,6 +117,7 @@ public:
         return compare(s, strlen(s));
     }
 
+    using base_type::ends_with;
     using base_type::starts_with;
 
     // compare to a C-style string
@@ -125,28 +126,10 @@ public:
         return helper::starts_with(*this, compare_to);
     }
 
-    ESTD_CPP_CONSTEXPR(14) bool starts_with(value_type ch) const
-    {
-        const bool r = *base_type::clock(0, 1) == ch;
-
-        base_type::cunlock();
-
-        return r;
-    }
-
     // compare to a C-style string
     constexpr bool ends_with(const_pointer compare_to) const
     {
         return helper::ends_with(*this, compare_to);
-    }
-
-    ESTD_CPP_CONSTEXPR(14) bool ends_with(value_type ch) const
-    {
-        const bool r = *base_type::clock(base_type::size() - 1, 1) == ch;
-
-        base_type::cunlock();
-
-        return r;
     }
 
     size_type find(value_type ch, size_type pos = 0) const